requirements.txt
click==6.7
dj-database-url==0.4.2
Django==1.11.7
djangorestframework==3.7.3
Flask==0.12.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.12
psycopg2==2.7.3.2
pytz==2017.3
Werkzeug==0.12.2
whitenoise==3.3.1
